% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CertificatBigData2020_Calcul_Statistique_Solution.R
\name{vCertificatBigData}
\alias{vCertificatBigData}
\title{fonction qui sert à faire des statistique sur le risque en entreprise}
\usage{
vCertificatBigData(
  n,
  paramX1_1 = 0,
  paramX1_2 = 1,
  paramX2_1 = 0,
  paramX2_2 = 1,
  loi = "uniforme",
  operation = "somme",
  graphics = FALSE
)
}
\arguments{
\item{n}{a number vkfjedgofih.}

\item{paramX1_1}{a number.}

\item{paramX1_2}{a number.}

\item{paramX2_1}{a number.}

\item{paramX2_2}{a number.}

\item{loi}{a number.}

\item{operation}{a number.}

\item{graphics}{a number.}
}
\value{
a number.
}
\description{
fonction qui sert à faire des statistique sur le risque en entreprise
}
\examples{
vCertificatBigData(10,0,1,0,1)
}
